bots_supervisor/requirements.txt
grpcio==1.51.3 
grpcio-tools==1.51.3
kubernetes==26.1.0
